feat(linter): adds more config to sort and organize imports better
