+++ boilerplate/app/components/Header.tsx
@@ -8,11 +8,12 @@ import {
   ViewStyle,
 } from "react-native"
 
-import { isRTL, translate } from "@/i18n"
-import { $styles } from "@/theme/styles"
-import { ExtendedEdge, useSafeAreaInsetsStyle } from "@/utils/useSafeAreaInsetsStyle"
-import { useAppTheme } from "@/utils/useAppTheme"
+import { isRTL } from "@/i18n"
+import { translate } from "@/i18n/translate"
 import type { ThemedStyle } from "@/theme"
+import { $styles } from "@/theme/styles"
+import { useAppTheme } from "@/utils/useAppTheme"
+import { ExtendedEdge, useSafeAreaInsetsStyle } from "@/utils/useSafeAreaInsetsStyle"
 
 import { IconTypes, PressableIcon } from "./Icon"
 import { Text, TextProps } from "./Text"
