+++ boilerplate/app/components/Screen.tsx
@@ -1,5 +1,3 @@
-import { useScrollToTop } from "@react-navigation/native"
-import { StatusBar, StatusBarProps, StatusBarStyle } from "expo-status-bar"
 import { ReactNode, useRef, useState } from "react"
 import {
   KeyboardAvoidingView,
@@ -12,11 +10,13 @@ import {
   View,
   ViewStyle,
 } from "react-native"
+import { StatusBar, StatusBarProps, StatusBarStyle } from "expo-status-bar"
+import { useScrollToTop } from "@react-navigation/native"
 import { KeyboardAwareScrollView } from "react-native-keyboard-controller"
 
 import { $styles } from "@/theme/styles"
-import { ExtendedEdge, useSafeAreaInsetsStyle } from "@/utils/useSafeAreaInsetsStyle"
 import { useAppTheme } from "@/utils/useAppTheme"
+import { ExtendedEdge, useSafeAreaInsetsStyle } from "@/utils/useSafeAreaInsetsStyle"
 
 export const DEFAULT_BOTTOM_OFFSET = 50
