+++ boilerplate/app/config/index.ts
@@ -14,8 +14,8 @@
  * Read more here: https://reactnative.dev/docs/security#storing-sensitive-info
  */
 import BaseConfig from "./config.base"
-import ProdConfig from "./config.prod"
 import DevConfig from "./config.dev"
+import ProdConfig from "./config.prod"
 
 let ExtraConfig = ProdConfig
