+++ boilerplate/app/screens/DemoCommunityScreen.tsx
@@ -4,17 +4,17 @@ import { Image, ImageStyle, TextStyle, View, ViewStyle } from "react-native"
 import { ListItem } from "@/components/ListItem"
 import { Screen } from "@/components/Screen"
 import { Text } from "@/components/Text"
+import { isRTL } from "@/i18n"
 import { DemoTabScreenProps } from "@/navigators/DemoNavigator"
+import type { ThemedStyle } from "@/theme"
 import { $styles } from "@/theme/styles"
 import { openLinkInBrowser } from "@/utils/openLinkInBrowser"
-import { isRTL } from "@/i18n"
-import type { ThemedStyle } from "@/theme"
 import { useAppTheme } from "@/utils/useAppTheme"
 
 const chainReactLogo = require("@assets/images/demo/cr-logo.png")
 const reactNativeLiveLogo = require("@assets/images/demo/rnl-logo.png")
-const reactNativeRadioLogo = require("@assets/images/demo/rnr-logo.png")
 const reactNativeNewsletterLogo = require("@assets/images/demo/rnn-logo.png")
+const reactNativeRadioLogo = require("@assets/images/demo/rnr-logo.png")
 
 export const DemoCommunityScreen: FC<DemoTabScreenProps<"DemoCommunity">> =
   function DemoCommunityScreen(_props) {
