+++ boilerplate/app/screens/DemoPodcastListScreen.tsx
@@ -28,9 +28,9 @@ import { ListView } from "@/components/ListView"
 import { Screen } from "@/components/Screen"
 import { Text } from "@/components/Text"
 import { Switch } from "@/components/Toggle/Switch"
+import { useEpisodes, Episode, useEpisode } from "@/context/EpisodeContext"
 import { isRTL } from "@/i18n"
 import { translate } from "@/i18n/translate"
-import { useEpisodes, Episode, useEpisode } from "@/context/EpisodeContext"
 import { DemoTabScreenProps } from "@/navigators/DemoNavigator"
 import type { ThemedStyle } from "@/theme"
 import { $styles } from "@/theme/styles"
