+++ boilerplate/app/screens/DemoShowroomScreen/DemoShowroomScreen.tsx
@@ -1,8 +1,8 @@
-import { Link, RouteProp, useRoute } from "@react-navigation/native"
 import { FC, ReactElement, useCallback, useEffect, useRef, useState } from "react"
 import { Image, ImageStyle, Platform, SectionList, TextStyle, View, ViewStyle } from "react-native"
-import { Drawer } from "react-native-drawer-layout"
+import { Link, RouteProp, useRoute } from "@react-navigation/native"
 import { type ContentStyle } from "@shopify/flash-list"
+import { Drawer } from "react-native-drawer-layout"
 
 import { ListItem } from "@/components/ListItem"
 import { ListView, type ListViewRef } from "@/components/ListView"
@@ -13,9 +13,9 @@ import { translate } from "@/i18n/translate"
 import { DemoTabParamList, DemoTabScreenProps } from "@/navigators/DemoNavigator"
 import type { Theme, ThemedStyle } from "@/theme"
 import { $styles } from "@/theme/styles"
-import { useSafeAreaInsetsStyle } from "@/utils/useSafeAreaInsetsStyle"
 import { hasValidStringProp } from "@/utils/hasValidStringProp"
 import { useAppTheme } from "@/utils/useAppTheme"
+import { useSafeAreaInsetsStyle } from "@/utils/useSafeAreaInsetsStyle"
 
 import * as Demos from "./demos"
 import { DrawerIconButton } from "./DrawerIconButton"
