+++ boilerplate/app/screens/DemoShowroomScreen/DrawerIconButton.tsx
@@ -1,6 +1,6 @@
 import { Pressable, PressableProps, ViewStyle, Platform } from "react-native"
-import Animated, { interpolate, interpolateColor, useAnimatedStyle } from "react-native-reanimated"
 import { useDrawerProgress } from "react-native-drawer-layout"
+import Animated, { interpolate, interpolateColor, useAnimatedStyle } from "react-native-reanimated"
 
 import { isRTL } from "@/i18n"
 import { useAppTheme } from "@/utils/useAppTheme"
