+++ boilerplate/app/screens/DemoShowroomScreen/demos/DemoAutoImage.tsx
@@ -3,12 +3,12 @@ import { Image, ImageStyle, TextStyle, View, ViewStyle } from "react-native"
 
 import { AutoImage } from "@/components/AutoImage"
 import { Text } from "@/components/Text"
+import { translate } from "@/i18n/translate"
 import type { ThemedStyle } from "@/theme"
 import { $styles } from "@/theme/styles"
-import { translate } from "@/i18n/translate"
 
-import { Demo } from "../DemoShowroomScreen"
 import { DemoDivider } from "../DemoDivider"
+import { Demo } from "../DemoShowroomScreen"
 import { DemoUseCase } from "../DemoUseCase"
 
 const $imageContainer: ViewStyle = {
