+++ boilerplate/app/screens/DemoShowroomScreen/demos/DemoCard.tsx
@@ -4,8 +4,8 @@ import { Button } from "@/components/Button"
 import { Card } from "@/components/Card"
 import { Icon } from "@/components/Icon"
 
-import { Demo } from "../DemoShowroomScreen"
 import { DemoDivider } from "../DemoDivider"
+import { Demo } from "../DemoShowroomScreen"
 import { DemoUseCase } from "../DemoUseCase"
 
 export const DemoCard: Demo = {
