+++ boilerplate/app/screens/LoginScreen.tsx
@@ -7,10 +7,10 @@ import { PressableIcon } from "@/components/Icon"
 import { Screen } from "@/components/Screen"
 import { Text } from "@/components/Text"
 import { TextField, type TextFieldAccessoryProps } from "@/components/TextField"
+import { useAuth } from "@/context/AuthContext"
 import type { AppStackScreenProps } from "@/navigators/AppNavigator"
 import type { ThemedStyle } from "@/theme"
 import { useAppTheme } from "@/utils/useAppTheme"
-import { useAuth } from "@/context/AuthContext"
 
 interface LoginScreenProps extends AppStackScreenProps<"Login"> {}
