+++ boilerplate/app/utils/useAppTheme.ts
@@ -1,7 +1,7 @@
 import { createContext, useCallback, useContext, useEffect, useMemo, useState } from "react"
 import { StyleProp, useColorScheme } from "react-native"
-import { DarkTheme, DefaultTheme, useTheme as useNavTheme } from "@react-navigation/native"
 import * as SystemUI from "expo-system-ui"
+import { DarkTheme, DefaultTheme, useTheme as useNavTheme } from "@react-navigation/native"
 
 import {
   type Theme,
