+++ docs/boilerplate/app/theme/Theming.md
@@ -147,7 +147,8 @@ return <NavigationContainer theme={navigationTheme} {...props} />
 There are many component libraries that offer light/dark modes to their components. Here's an example of how to use `react-native-elements` with Ignite's theming system by extending their own `ThemeProvider`:
 
 ```tsx
-import { colorsDark, colorsLight, customFontsToLoad } from "@/theme"
+import { colorsDark, colorsLight } from "@/theme/colors"
+import { customFontsToLoad } from "@/theme/typography"
 import { createTheme as createRNEUITheme, ThemeProvider as RNEUIThemeProvider } from "@rneui/themed"
 
 export const ThemedRNEUIProvider = ({ children }) => {
